requirements.txt
streamlit>=1.30.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pandas>=2.1.0
numpy>=1.25.0
plotly>=5.18.0
openpyxl>=3.1.0
reportlab>=4.0.0
openai>=1.10.0
google-generativeai>=0.3.0
requests>=2.31.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
